// src/app/utils/Enums/node-headers.ts
export enum NodeHeaderTypes{
  common_point = "Common-point",
  Risk ="Risk",
  Control = "Control",
  Cause = "Cause",  
  Consequence = "Consequence",
  Linked_Risk ="Linked Risk",
  Hierarchy_Linkages = "Hierarchy Linkages",
  Kpi ="KPI",
  Incident =  "Incident",
  Audit = "Audit",
  Audit_Finding = "Audit Finding",
  Audit_Recommendations = "Audit Recommendations",
  Obligation = "Obligation",
  Policy = "Policy",
  Authority_Document = "Authority Document",
  Treatment = "Action"
 
}